Plutarch.Internal.Eq module, move PEq over
