+++ Plutarch/BitString.hs
@@ -13,11 +13,12 @@ module Plutarch.BitString (
 ) where
 
 import GHC.Generics (Generic)
-import Plutarch.Bool (PEq, POrd, PPartialOrd, pif, (#<))
+import Plutarch.Bool (POrd, PPartialOrd, pif, (#<))
 import Plutarch.Builtin (PBuiltinList)
 import Plutarch.Builtin.Bool (PBool)
 import Plutarch.ByteString (PByteString)
 import Plutarch.Integer (PInteger)
+import Plutarch.Internal.Eq (PEq)
 import Plutarch.Internal.Newtype (PlutusTypeNewtype)
 import Plutarch.Internal.Other (pto)
 import Plutarch.Internal.PLam (plam)
