+++ Plutarch/Bool.hs
@@ -2,7 +2,6 @@
 {-# OPTIONS_GHC -Wno-orphans #-}
 
 module Plutarch.Bool (
-  PEq (..),
   PPartialOrd (..),
   POrd (..),
   pif,
@@ -27,20 +26,8 @@ module Plutarch.Bool (
   psor',
 ) where
 
-import Data.List.NonEmpty (nonEmpty)
-import Generics.SOP (
-  All,
-  All2,
-  HCollapse (hcollapse),
-  K (K),
-  NP,
-  Proxy (Proxy),
-  SOP (SOP),
-  ccompare_NS,
-  hcliftA2,
- )
 import Plutarch.Builtin.Bool (PBool (PFalse, PTrue))
-import Plutarch.Internal.Generic (PCode, PGeneric, gpfrom)
+import Plutarch.Internal.Eq (PEq ((#==)))
 import Plutarch.Internal.Lift (pconstant)
 import Plutarch.Internal.Other (
   pto,
@@ -63,17 +50,6 @@ import Plutarch.Internal.Term (
  )
 import Plutarch.Unsafe (punsafeBuiltin)
 import PlutusCore qualified as PLC
-
-class PEq t where
-  (#==) :: Term s t -> Term s t -> Term s PBool
-  default (#==) ::
-    (PGeneric t, PlutusType t, All2 PEq (PCode t)) =>
-    Term s t ->
-    Term s t ->
-    Term s PBool
-  a #== b = gpeq # a # b
-
-infix 4 #==
 
 {- | Partial ordering relation.
 
@@ -206,40 +182,6 @@ por = phoistAcyclic $ plam $ \x -> pif' # x # phoistAcyclic (pdelay $ pcon PTrue
 -- | Hoisted, Plutarch level, strictly evaluated boolean or function.
 por' :: Term s (PBool :--> PBool :--> PBool)
 por' = phoistAcyclic $ plam $ \x -> pif' # x # pcon PTrue
-
--- | Like Haskell's `and` but for Plutarch terms
-pands :: [Term s PBool] -> Term s PBool
-pands ts' =
-  case nonEmpty ts' of
-    Nothing -> pcon PTrue
-    Just ts -> foldl1 (#&&) ts
-
--- | Generic version of (#==)
-gpeq ::
-  forall t s.
-  ( PGeneric t
-  , PlutusType t
-  , All2 PEq (PCode t)
-  ) =>
-  Term s (t :--> t :--> PBool)
-gpeq =
-  phoistAcyclic $
-    plam $ \x y ->
-      pmatch x $ \x' ->
-        pmatch y $ \y' ->
-          gpeq' (gpfrom x') (gpfrom y')
-
-gpeq' :: All2 PEq xss => SOP (Term s) xss -> SOP (Term s) xss -> Term s PBool
-gpeq' (SOP c1) (SOP c2) =
-  ccompare_NS (Proxy @(All PEq)) (pcon PFalse) eqProd (pcon PFalse) c1 c2
-
-eqProd :: All PEq xs => NP (Term s) xs -> NP (Term s) xs -> Term s PBool
-eqProd p1 p2 =
-  pands $ hcollapse $ hcliftA2 (Proxy :: Proxy PEq) eqTerm p1 p2
-  where
-    eqTerm :: forall s a. PEq a => Term s a -> Term s a -> K (Term s PBool) a
-    eqTerm a b =
-      K $ a #== b
 
 -- | 'PInner' of 'PSBool'.
 newtype PSBoolRaw (a :: PType) (s :: S) = PSBoolRaw (Term s (a :--> a :--> a))
