+++ Plutarch/Show.hs
@@ -27,7 +27,7 @@ import Generics.SOP (
   hmap,
  )
 import Generics.SOP.GGP (gdatatypeInfo)
-import Plutarch.Bool (PEq, pif, pif', (#<), (#==))
+import Plutarch.Bool (pif, pif', (#<))
 import Plutarch.Builtin.Bool (PBool)
 import Plutarch.ByteString (
   PByte,
@@ -40,6 +40,7 @@ import Plutarch.ByteString (
   punsafeIntegerToByte,
  )
 import Plutarch.Integer (PInteger, PIntegral (pquot, prem))
+import Plutarch.Internal.Eq (PEq ((#==)))
 import Plutarch.Internal.Generic (PCode, PGeneric, gpfrom)
 import Plutarch.Internal.Lift (pconstant)
 import Plutarch.Internal.Other (
