+++ Plutarch/String.hs
@@ -16,10 +16,11 @@ import Data.String (IsString, fromString)
 import Data.Text (Text)
 import Data.Text qualified as Text
 import GHC.Generics (Generic)
-import Plutarch.Bool (PEq, (#&&), (#<=), (#==), (#||))
+import Plutarch.Bool ((#&&), (#<=), (#||))
 import Plutarch.Builtin.Bool (PBool)
 import Plutarch.ByteString (PByteString)
 import Plutarch.Integer (PInteger)
+import Plutarch.Internal.Eq (PEq ((#==)))
 import Plutarch.Internal.Lift (DeriveBuiltinPLiftable, PLiftable, PLifted (PLifted), pconstant)
 import Plutarch.Internal.Newtype (PlutusTypeNewtype)
 import Plutarch.Internal.Other (POpaque)
